requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
python-dotenv==1.1.1

# SQLAlchemy and Database
alembic==1.16.2
SQLAlchemy==2.0.41
PyMySQL

# Pandas e processamento de arquivos
pandas==2.3.0
openpyxl==3.1.5

# Outras dependências
blinker==1.9.0
click==8.2.1
colorama==0.4.6
et-xmlfile==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.1
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing-extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
greenlet==3.2.3